+++ src/Module/Infrastructure/ModuleListInfrastructure.php
@@ -10,16 +10,18 @@ declare(strict_types=1);
 namespace OxidEsales\GraphQL\ConfigurationAccess\Module\Infrastructure;
 
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Bridge\ShopConfigurationDaoBridgeInterface;
+use OxidEsales\GraphQL\ConfigurationAccess\Module\DataType\ModuleDataTypeFactoryInterface;
 use OxidEsales\GraphQL\ConfigurationAccess\Module\Exception\ModulesNotFoundException;
 
 class ModuleListInfrastructure implements ModuleListInfrastructureInterface
 {
     public function __construct(
-        private readonly ShopConfigurationDaoBridgeInterface $shopConfigurationDaoBridge
+        private readonly ShopConfigurationDaoBridgeInterface $shopConfigurationDaoBridge,
+        private readonly ModuleDataTypeFactoryInterface $moduleDataTypeFactory
     ) {
     }
 
-    public function getModuleConfigurations(): array
+    public function getModuleList(): array
     {
         $shopConfiguration = $this->shopConfigurationDaoBridge->get();
         $moduleConfigurations = $shopConfiguration->getModuleConfigurations();
@@ -28,6 +30,10 @@ class ModuleListInfrastructure implements ModuleListInfrastructureInterface
             throw new ModulesNotFoundException();
         }
 
-        return $moduleConfigurations;
+        $moduleDatatypes = [];
+        foreach ($moduleConfigurations as $moduleConfig) {
+            $moduleDatatypes[] = $this->moduleDataTypeFactory->createFromModuleConfiguration($moduleConfig);
+        }
+        return $moduleDatatypes;
     }
 }