+++ tests/Unit/Shared/Exception/CollectionEncodingExceptionTest.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\GraphQL\ConfigurationAccess\Tests\Unit\Shared\Exception;
+
+use OxidEsales\GraphQL\ConfigurationAccess\Shared\Exception\CollectionEncodingException;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \OxidEsales\GraphQL\ConfigurationAccess\Shared\Exception\CollectionEncodingException
+ */
+class CollectionEncodingExceptionTest extends TestCase
+{
+    public function testExceptionMessage()
+    {
+        $sut = new CollectionEncodingException();
+        $this->assertSame('Error encountered while encoding collection data', $sut->getMessage());
+    }
+}