+++ tests/Unit/Theme/Exception/NoSettingsFoundForThemeExceptionTest.php
@@ -7,13 +7,14 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\GraphQL\ConfigurationAccess\Tests\Unit\Shared\Exception;
+namespace OxidEsales\GraphQL\ConfigurationAccess\Tests\Unit\Theme\Exception;
 
 use OxidEsales\GraphQL\ConfigurationAccess\Theme\Exception\NoSettingsFoundForThemeException;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @covers \OxidEsales\GraphQL\ConfigurationAccess\Theme\Exception\NoSettingsFoundForThemeException
+ * @covers \OxidEsales\GraphQL\ConfigurationAccess\Shared\Exception\NoSettingsFoundException
  */
 class NoSettingsFoundForThemeExceptionTest extends TestCase
 {